Move ContextExtension serde impl to ergotree-ir
