+++ ergo-lib/Cargo.toml
@@ -6,9 +6,7 @@ authors = ["Denys Zadorozhnyi <denys@zadorozhnyi.com>"]
 repository.workspace = true
 edition.workspace = true
 description = "ErgoTree interpreter and wallet-like features for Ergo"
-exclude = [
-    "proptest-regressions/*"
-]
+exclude = ["proptest-regressions/*"]
 
 [lib]
 crate-type = ["cdylib", "rlib"]
@@ -20,9 +18,9 @@ ergo-chain-types = { workspace = true }
 ergotree-ir = { workspace = true }
 ergotree-interpreter = { workspace = true }
 ergo-nipopow = { workspace = true }
-ergoscript-compiler = { workspace = true, optional = true}
+ergoscript-compiler = { workspace = true, optional = true }
 ergo-merkle-tree = { workspace = true }
-ergo-rest = { workspace = true, optional = true}
+ergo-rest = { workspace = true, optional = true }
 indexmap = { workspace = true }
 base16 = { workspace = true }
 serde = { workspace = true, optional = true }
@@ -31,7 +29,7 @@ thiserror = { workspace = true }
 derive_more = { workspace = true }
 bounded-vec = { workspace = true }
 num-bigint = { workspace = true, features = ["serde"] }
-proptest-derive = {workspace = true, optional = true }
+proptest-derive = { workspace = true, optional = true }
 k256 = { workspace = true }
 sha2 = { workspace = true }
 hmac = { version = "0.12" }
@@ -40,16 +38,28 @@ rand = { workspace = true }
 bitvec = { workspace = true, optional = true }
 unicode-normalization = "0.1.19"
 lazy_static = { workspace = true }
-proptest = { workspace = true , optional = true }
+proptest = { workspace = true, optional = true }
 serde_with = { workspace = true, optional = true }
 itertools = { workspace = true }
 
 
 [features]
 default = ["json"]
-json = ["serde", "serde_json", "serde_with", "bounded-vec/serde"]
+json = [
+    "serde",
+    "serde_json",
+    "serde_with",
+    "bounded-vec/serde",
+    "ergotree-ir/json",
+]
 compiler = ["ergoscript-compiler"]
-arbitrary = ["proptest", "proptest-derive", "ergotree-ir/arbitrary", "ergo-chain-types/arbitrary", "ergotree-interpreter/arbitrary"]
+arbitrary = [
+    "proptest",
+    "proptest-derive",
+    "ergotree-ir/arbitrary",
+    "ergo-chain-types/arbitrary",
+    "ergotree-interpreter/arbitrary",
+]
 mnemonic_gen = ["bitvec"]
 rest = ["ergo-rest"]
