+++ ergo-lib/src/chain/json.rs
@@ -4,7 +4,6 @@ use serde::{Deserialize, Serialize};
 
 use ergotree_interpreter::sigma_protocol::prover::ProofBytes;
 
-pub(crate) mod context_extension;
 pub(crate) mod hint;
 pub(crate) mod parameters;
 pub(crate) mod transaction;
