+++ ergotree-ir/src/chain/json.rs
@@ -4,7 +4,6 @@ use core::fmt;
 use core::marker::PhantomData;
 use core::str::FromStr;
 
-
 use serde::de;
 use serde::de::MapAccess;
 use serde::de::Visitor;
